Added fetching toml for node config

+++ desktop/nodeManager.js
@@ -1,10 +1,17 @@
 import path from 'path';
 import os from 'os';
+import fs from 'fs';
+import util from 'util';
 import { app } from 'electron';
 import { ipcConsts } from '../app/vars';
+import FileManager from './fileManager';
 
 const { execFile, exec } = require('child_process');
+const fetch = require('node-fetch');
+const toml = require('toml');
 const find = require('find-process');
+
+const writeFileAsync = util.promisify(fs.writeFile);
 
 const osTargetNames = {
   Darwin: 'mac',
@@ -55,29 +62,50 @@ class NodeManager {
     }
   };
 
-  static tmpRunNodeFunc = ({ port }) => {
-    const homeDirPath = app.getPath('home');
-    const dataPath = path.resolve(homeDirPath, 'spacemesh');
-    const testDataPath = path.resolve(homeDirPath, 'spacemeshtestdata');
-    const postDataPath = path.resolve(homeDirPath, 'post');
-    const logFilePath = path.resolve(app.getPath('documents'), 'spacemeshLog.txt');
-    const command =
-      os.type() === 'windows'
-        ? `rmdir /q/s ${dataPath} && rmdir /q/s ${testDataPath} && rmdir /q/s ${postDataPath} && del ${logFilePath}`
-        : `rm -rf ${dataPath} && rm -rf ${testDataPath} && rm -rf ${postDataPath} && rm -rf ${logFilePath}`;
-    exec(command, (err) => {
-      if (!err) {
-        const pathWithParams = `./go-spacemesh --grpc-server --json-server --tcp-port ${port} --poet-server spacemesh-testnet-poet-grpc-lb-949d0cde858743fb.elb.us-east-1.amazonaws.com:50002 --post-space 8589934592 --test-mode --randcon 8 --layer-duration-sec 180 --hare-wakeup-delta 30 --hare-round-duration-sec 30 --layers-per-epoch 480 --eligibility-confidence-param 200 --eligibility-epoch-offset 0 --layer-average-size 50 --genesis-active-size 300 --hare-committee-size 50 --hare-max-adversaries 24 --sync-request-timeout 60000 --post-labels 100 --max-inbound 12 --genesis-time 2020-01-08T21:31:19+00:00 --bootstrap --bootnodes spacemesh://3augAJDbLGyXGfGxQT3GvTPaBdvbDLx8tyH7vZrGnRJT@54.180.96.161:64587 -d ~/spacemeshtestdata/ > ${logFilePath}`;
+  static tmpRunNodeFunc = async ({ port, store }) => {
+    const rawData = await fetch('http://nodes.unruly.io');
+    const tomlData = await rawData.text();
+    try {
+      const parsedToml = toml.parse(tomlData);
+      const fetchedGenesisTime = parsedToml.main['genesis-time'];
+      const prevGenesisTime = store.get('genesisTime') || '';
+      const homeDirPath = app.getPath('home');
+      const dataPath = path.resolve(homeDirPath, 'spacemesh');
+      const testDataPath = path.resolve(homeDirPath, 'spacemeshtestdata');
+      const postDataPath = path.resolve(homeDirPath, 'post');
+      const logFilePath = path.resolve(app.getPath('documents'), 'spacemeshLog.txt');
+      const pathWithParams = `./go-spacemesh --grpc-server --json-server --tcp-port ${port} -d ~/spacemeshtestdata/ >> ${logFilePath}`;
+      await writeFileAsync('./config.toml', tomlData);
+      if (prevGenesisTime !== fetchedGenesisTime) {
+        store.set('genesisTime', fetchedGenesisTime);
+        await FileManager.cleanWalletFile();
+        const command =
+          os.type() === 'windows'
+            ? `rmdir /q/s ${dataPath} && rmdir /q/s ${testDataPath} && rmdir /q/s ${postDataPath} && del ${logFilePath}`
+            : `rm -rf ${dataPath} && rm -rf ${testDataPath} && rm -rf ${postDataPath} && rm -rf ${logFilePath}`;
+        exec(command, (err) => {
+          if (!err) {
+            exec(pathWithParams, (error) => {
+              if (error) {
+                console.error(error); // eslint-disable-line no-console
+              }
+              console.log('node started with provided params'); // eslint-disable-line no-console
+            });
+          } else {
+            console.error(err); // eslint-disable-line no-console
+          }
+        });
+      } else {
         exec(pathWithParams, (error) => {
           if (error) {
             console.error(error); // eslint-disable-line no-console
           }
           console.log('node started with provided params'); // eslint-disable-line no-console
         });
-      } else {
-        console.error(err); // eslint-disable-line no-console
       }
-    });
+    } catch (e) {
+      console.error(`Parsing error on line ${e.line}, column ${e.column}: ${e.message}`); // eslint-disable-line no-console
+    }
   };
 }
